Remove empty parts after force-split
Force-split created emtpy tokens between force-split characters, e.g. in words with two hyphens: h--a --> ['h', '-', '', '-','a']. This caused problems in the recursive splitting, in the end resulting in negative atom count for hyphens. The problems only reveal themselves in segmentation phase, as a log is taken of the atom count.